refactor: use Gate::resource

+++ app/Http/Controllers/BlogPostController.php
@@ -113,6 +113,10 @@ class BlogPostController extends Controller
     public function destroy($id)
     {
         $blogpost = BlogPost::findOrFail($id);
+        $this->authorize('blogposts.delete', $blogpost);
+        $comments = $blogpost->comments;
+        foreach($comments as $comment)
+            $comment->delete();
         $blogpost->delete();
 
         session()->flash('status','Blog Post with the ID' . $id .' was deleted!');
